Added ChainedRollable, ChainedTable, respective hooks, builders and hook builders
The idea of the ChainedTable comes from a voiced mild frustration of chaining and giga-indenting weighted tables or chance tables to get the same effect, but a lot uglier.

It took me a bit to get the vibe down, but with feedback I got something much preferred to the other options that were discussed and over the previous "attempts".

The example I have implemented is the Yama drop table, and it can be adjusted using the hooks to adjust for duo/contribution/etc, though I don't know those exact calculations. It's probably simple. This also enables a "proper" wintertodt reward cart pull.